Fix typing issues.
